More PR cleanup and some refactoring
