+++ zoned-datetime-picker/package.json
@@ -3,7 +3,7 @@
   "description": "Zoned DateTime Picker is a DatoCMS plugin that stores datetime timestamps with time zone information. Outputs a JSON with ISO8601 and IXDTF strings, the IANA time zone string, UTC offset, and more.",
   "homepage": "https://github.com/datocms/plugins/tree/master/zoned-datetime-picker",
   "private": false,
-  "version": "0.2.0",
+  "version": "0.1.8",
   "author": "DatoCMS <support@datocms.com>",
   "type": "module",
   "keywords": [
